implemented a watchdog timer.
The WDT monitors the server log. If no log statements are printed for 100s it issues an invalid command. This forces the server to write a few lines to the log if the server is running but simply has no activity. If the server fails to respond for 300s (5 min) it is assumed to be frozen and is forcibly killed. This will also kill the proxy. If so configured, the docker service will simply restart the container after it exits. This servers as a simple restart mechanism.